# Conflicts:
#	TMessagesProj/build.gradle
#	TMessagesProj/config/release/AndroidManifest_standalone.xml
#	TMessagesProj/src/main/java/org/telegram/messenger/BuildVars.java
#	gradle.properties

+++ TMessagesProj/src/main/java/org/telegram/ui/Components/RLottieDrawable.java
@@ -21,7 +21,6 @@ import android.os.Build;
 import android.os.Handler;
 import android.os.Looper;
 import android.text.TextUtils;
-import android.util.Log;
 import android.view.HapticFeedbackConstants;
 import android.view.View;
 
@@ -109,8 +108,8 @@ public class RLottieDrawable extends BitmapDrawable implements Animatable, Bitma
     private boolean applyTransformation;
     private boolean needScale;
     private final RectF dstRect = new RectF();
-    private RectF dstRectBackground;
-    private Paint backgroundPaint;
+    private RectF[] dstRectBackground = new RectF[DrawingInBackgroundThreadDrawable.THREAD_COUNT];
+    private Paint[] backgroundPaint = new Paint[DrawingInBackgroundThreadDrawable.THREAD_COUNT];
     protected static final Handler uiHandler = new Handler(Looper.getMainLooper());
     protected volatile boolean isRunning;
     protected volatile boolean isRecycled;
@@ -176,6 +175,7 @@ public class RLottieDrawable extends BitmapDrawable implements Animatable, Bitma
                 if (lottieCacheGenerateQueue == null) {
                     createCacheGenQueue();
                 }
+                BitmapsCache.incrementTaskCounter();
                 lottieCacheGenerateQueue.postRunnable(cacheGenerateTask = () -> {
                     BitmapsCache bitmapsCacheFinal = bitmapsCache;
                     if (bitmapsCacheFinal != null) {
@@ -190,7 +190,10 @@ public class RLottieDrawable extends BitmapDrawable implements Animatable, Bitma
     private Runnable uiRunnableCacheFinished = new Runnable() {
         @Override
         public void run() {
-            cacheGenerateTask = null;
+            if (cacheGenerateTask != null) {
+                BitmapsCache.decrementTaskCounter();
+                cacheGenerateTask = null;
+            }
             generatingCache = false;
             decodeFrameFinishedInternal();
         }
@@ -206,6 +209,7 @@ public class RLottieDrawable extends BitmapDrawable implements Animatable, Bitma
     protected void checkRunningTasks() {
         if (cacheGenerateTask != null) {
             lottieCacheGenerateQueue.cancelRunnable(cacheGenerateTask);
+            BitmapsCache.decrementTaskCounter();
             cacheGenerateTask = null;
         }
         if (!hasParentView() && nextRenderingBitmap != null && loadFrameTask != null) {
@@ -741,6 +745,7 @@ public class RLottieDrawable extends BitmapDrawable implements Animatable, Bitma
                 if (bitmapsCache != null) {
                     bitmapsCache.cancelCreate();
                 }
+                BitmapsCache.decrementTaskCounter();
             }, 600);
         } else if (!mustCancel && cancelCache != null) {
             AndroidUtilities.cancelRunOnUIThread(cancelCache);
@@ -906,10 +911,6 @@ public class RLottieDrawable extends BitmapDrawable implements Animatable, Bitma
     }
 
     protected boolean scheduleNextGetFrame() {
-        return scheduleNextGetFrame(false);
-    }
-
-    protected boolean scheduleNextGetFrame(boolean allowGroupedUpdateLocal) {
         if (loadFrameTask != null || nextRenderingBitmap != null || !canLoadFrames() || loadingInBackground || destroyWhenDone || !isRunning && (!decodeSingleFrame || decodeSingleFrame && singleFrameDecoded)) {
             return false;
         }
@@ -925,8 +926,8 @@ public class RLottieDrawable extends BitmapDrawable implements Animatable, Bitma
             newReplaceColors = null;
         }
         loadFrameTask = loadFrameRunnable;
-        if (allowGroupedUpdateLocal && shouldLimitFps) {
-            DispatchQueuePoolBackground.execute(loadFrameTask);
+        if (shouldLimitFps) {
+            DispatchQueuePoolBackground.execute(loadFrameTask, frameWaitSync != null);
         } else {
             loadFrameRunnableQueue.execute(loadFrameTask);
         }
@@ -977,7 +978,7 @@ public class RLottieDrawable extends BitmapDrawable implements Animatable, Bitma
         if (resetFrame && !isRunning) {
             isRunning = true;
         }
-        if (scheduleNextGetFrame(false)) {
+        if (scheduleNextGetFrame()) {
             if (!async) {
                 try {
                     frameWaitSync.await();
@@ -1081,27 +1082,27 @@ public class RLottieDrawable extends BitmapDrawable implements Animatable, Bitma
                 }
             }
         }
-        scheduleNextGetFrame(true);
+        scheduleNextGetFrame();
     }
 
     @Override
     public void draw(Canvas canvas) {
-        drawInternal(canvas, false, 0);
+        drawInternal(canvas, false, 0, 0);
     }
 
-    public void drawInBackground(Canvas canvas, float x, float y, float w, float h, int alpha, ColorFilter colorFilter) {
-        if (dstRectBackground == null) {
-            dstRectBackground = new RectF();
-            backgroundPaint = new Paint(Paint.ANTI_ALIAS_FLAG);
-            backgroundPaint.setFilterBitmap(true);
+    public void drawInBackground(Canvas canvas, float x, float y, float w, float h, int alpha, ColorFilter colorFilter, int threadIndex) {
+        if (dstRectBackground[threadIndex] == null) {
+            dstRectBackground[threadIndex] = new RectF();
+            backgroundPaint[threadIndex] = new Paint(Paint.ANTI_ALIAS_FLAG);
+            backgroundPaint[threadIndex].setFilterBitmap(true);
         }
-        backgroundPaint.setAlpha(alpha);
-        backgroundPaint.setColorFilter(colorFilter);
-        dstRectBackground.set(x, y, x + w, y + h);
-        drawInternal(canvas, true, 0);
+        backgroundPaint[threadIndex].setAlpha(alpha);
+        backgroundPaint[threadIndex].setColorFilter(colorFilter);
+        dstRectBackground[threadIndex].set(x, y, x + w, y + h);
+        drawInternal(canvas, true, 0, threadIndex);
     }
 
-    public void drawInternal(Canvas canvas, boolean drawInBackground, long time) {
+    public void drawInternal(Canvas canvas, boolean drawInBackground, long time, int threadIndex) {
         if (!canLoadFrames() || destroyWhenDone) {
             return;
         }
@@ -1109,8 +1110,8 @@ public class RLottieDrawable extends BitmapDrawable implements Animatable, Bitma
             updateCurrentFrame(time, false);
         }
 
-        RectF rect = drawInBackground ? dstRectBackground : dstRect;
-        Paint paint = drawInBackground ? backgroundPaint : getPaint();
+        RectF rect = drawInBackground ? dstRectBackground[threadIndex] : dstRect;
+        Paint paint = drawInBackground ? backgroundPaint[threadIndex] : getPaint();
 
         if (paint.getAlpha() == 0) {
             return;
@@ -1164,7 +1165,7 @@ public class RLottieDrawable extends BitmapDrawable implements Animatable, Bitma
         }
         if (isRunning) {
             if (renderingBitmap == null && nextRenderingBitmap == null) {
-                scheduleNextGetFrame(true);
+                scheduleNextGetFrame();
             } else if (nextRenderingBitmap != null && (renderingBitmap == null || (timeDiff >= timeCheck && !skipFrameUpdate))) {
                 if (vibrationPattern != null && currentParentView != null && allowVibration) {
                     Integer force = vibrationPattern.get(currentFrame - 1);
@@ -1243,7 +1244,7 @@ public class RLottieDrawable extends BitmapDrawable implements Animatable, Bitma
     @Override
     public void prepareForGenerateCache() {
         generateCacheNativePtr = create(args.file.toString(), args.json, width, height, new int[3], false, args.colorReplacement, false, args.fitzModifier);
-        if (generateCacheNativePtr == 0) {
+        if (generateCacheNativePtr == 0 && file != null) {
             file.delete();
         }
     }
@@ -1323,13 +1324,22 @@ public class RLottieDrawable extends BitmapDrawable implements Animatable, Bitma
                     if (lottieCacheGenerateQueue == null) {
                         createCacheGenQueue();
                     }
-                    lottieCacheGenerateQueue.postRunnable(cacheGenerateTask = () -> {
-                        BitmapsCache bitmapsCacheFinal = bitmapsCache;
-                        if (bitmapsCacheFinal != null) {
-                            bitmapsCacheFinal.createCache();
-                        }
-                        AndroidUtilities.runOnUIThread(onReady);
-                    });
+                    if (cacheGenerateTask == null) {
+                        BitmapsCache.incrementTaskCounter();
+                        lottieCacheGenerateQueue.postRunnable(cacheGenerateTask = () -> {
+                            BitmapsCache bitmapsCacheFinal = bitmapsCache;
+                            if (bitmapsCacheFinal != null) {
+                                bitmapsCacheFinal.createCache();
+                            }
+                            AndroidUtilities.runOnUIThread(() -> {
+                                onReady.run();
+                                if (cacheGenerateTask != null) {
+                                    cacheGenerateTask = null;
+                                    BitmapsCache.decrementTaskCounter();
+                                }
+                            });
+                        });
+                    }
                 });
             }
         });
